requirements.txt
google-auth-oauthlib==1.0.0
google-auth==2.22.0
google-api-python-client==2.95.0
requests==2.31.0
python-dateutil==2.8.2